Added Department

+++ ministryofJusticeApp/ministryofjusticeWebUi/Models/DepartmentViewModel.cs
@@ -3,14 +3,14 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
+using ministryofjusticeDomain.Entities;
 
 namespace ministryofjusticeWebUi.Models
 {
     public class DepartmentViewModel
     {
-        public byte Id { get; set; }
+        public int Id { get; set; }
 
-        [Required]
-        public string DepartmentName { get; set; }
+        [Required] public string DepartmentName { get; set; }
     }
 }
